requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.24.0
openai==1.3.0
beautifulsoup4==4.12.2
pyperclip==1.8.2
selenium>=4.10.0
pandas>=2.0.3
numpy>=1.24.3
markdown==3.7
pymongo==4.11.1
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
python-markdown-math==0.8 